+++ history.md
@@ -37,6 +37,7 @@ node starsky-tools/build-tools/app-version-update.js
 ```
 # version 0.5.0-beta.7 _(Unreleased)_ - 2022-09-??
 - [x]   (Security) _Back-end_ Upgrade Nuget packages (PR #878)
+- [x]   (Security) _Back-end_ Update Security Headers e.g. CSP, Permissions Policy (PR #880)
 
 # version 0.5.0-beta.7 - 2022-09-20
 - [x]   (Changed) _Back-end_ Upgrade to .NET 6 - SDK 6.0.302 (Runtime: 6.0.7) _overwritten_ (PR #838)